requirements.txt
asgiref==3.5.0
boto3==1.21.34
botocore==1.24.34
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
Django==4.0.3
django-cors-headers==3.11.0
django-dotenv==1.4.2
django-filter==21.1
django-sslserver==0.22
django-storages==1.12.3
djangorestframework==3.13.1
idna==3.3
importlib-metadata==4.11.3
jmespath==1.0.0
Markdown==3.3.6
Pillow==9.1.0
pycparser==2.21
PyJWT==2.3.0
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
s3transfer==0.5.2
sentry-sdk==1.5.8
six==1.16.0
sonoff-python==0.2.1
sqlparse==0.4.2
urllib3==1.26.9
websocket-client==1.3.2
zipp==3.8.0
